// src/views/certificationReviewCenter/cbam/ReportForm/type.ts
export interface SaleProductResp {
  /**
   * 其他合金比例
   */
  alloyPer?: number;
  /**
   * 每吨铝使用废铝
   */
  alUse?: number;
  /**
   * 铵态氮占比
   */
  ammonium?: number;
  /**
   * 是否煅烧:0是;1否
   */
  calcine?: number;
  cbamId?: number;
  /**
   * 熟料
   */
  clinker?: number;
  /**
   * cn分类
   */
  cnCode?: string;
  cnName?: string;
  cper?: number;
  /**
   * 铬比例
   */
  crPer?: number;
  delFlag?: number;
  /**
   * 排放强度
   */
  emission?: number;
  /**
   * 1配置中;2已完成(1:配置中; 2:已完成)
   */
  fillStatus?: number;
  id?: number;
  /**
   * 其他材料占比
   */
  materialPer?: number;
  /**
   * 锰比例
   */
  mnPer?: number;
  /**
   * 镍比例
   */
  niPer?: number;
  /**
   * 硝酸比例
   */
  nitric?: number;
  /**
   * 氮元素
   */
  nitrogen?: number;
  /**
   * 非铝元素占比
   */
  nonAl?: number;
  /**
   * 硝酸盐占比
   */
  noPer?: number;
  /**
   * 有机占比
   */
  organic?: number;
  /**
   * 工序
   */
  processId?: number;
  processName?: string;
  /**
   * 产品名称
   */
  productName?: string;
  /**
   * 前体的主要还原剂
   */
  reducing?: number;
  /**
   * 含水溶液浓度
   */
  solution?: number;
  /**
   * 钢厂标识号
   */
  steelCode?: string;
  /**
   * 每吨使用废钢
   */
  steelScrap?: number;
  /**
   * 尿素
   */
  urea?: number;
  /**
   * 尿素占比
   */
  urPer?: number;
  /**
   * 消费前废料占比
   */
  wasteMaterial?: number;
  [property: string]: any;
}
// ...
export interface SunDTO {
  child?: SunDTO[];
  /**
   * 间接排放受影响的
   */
  inputPower?: number;
  /**
   * 0直接;1间接
   */
  isDefault?: number;
  /**
   * 0工序;1前体
   */
  isProcess?: number;
  /**
   * 排放受影响的
   */
  outPower?: number;
  /**
   * 百分比
   */
  per?: number;
  processId?: number;
  processName?: string;
  unitName?: string;
  [property: string]: any;
}